Convert main and plugins to C++
While libdrakvuf maintains its own state and thus multiple can be used in a single
program, the plugins have largely assumed that only one instance will be active
at any given time. While the plugin state could be maintained in pure C structs as well,
this adds considerably more work and is less intuitive then maintaing C++ classes.
