+++ src/libdrakvuf/Makefile.am
@@ -102,7 +102,7 @@
  #                                                                         #
  #*************************************************************************# 
 
-h_sources = drakvuf.h
+h_sources = libdrakvuf.h
 c_sources = drakvuf.c injector.c win-exports.c vmi.c win-symbols.c win-handles.c win-processes.c
 
 AM_CFLAGS = -I$(top_srcdir)
