+++ src/libdrakvuf/private.h
@@ -115,7 +115,7 @@
 #include <libvmi/libvmi.h>
 #include <libvmi/events.h>
 
-#include "drakvuf.h"
+#include "libdrakvuf.h"
 #include "../xen_helper/xen_helper.h"
 
 #ifdef DRAKVUF_DEBUG
