+++ src/plugins/exmon/exmon.h
@@ -105,20 +105,25 @@
 #ifndef EXMON_H
 #define EXMON_H
 
-#ifdef ENABLE_PLUGIN_EXMON
+#include "plugins/plugins.h"
 
-int plugin_exmon_start(drakvuf_t drakvuf, const void *config);
-int plugin_exmon_stop(drakvuf_t drakvuf);
+class exmon: public plugin {
+    public:
+        drakvuf_trap_t trap = {
+            .lookup_type = LOOKUP_PID,
+            .u.pid = 4,
+            .addr_type = ADDR_RVA,
+            .name = "KiDispatchException",
+            .module = "ntoskrnl.exe",
+            .type = BREAKPOINT
+        };
+        output_format_t format;
+        page_mode_t pm;
+        size_t *offsets;
+        size_t ktrap_frame_size;
 
-#else
-
-static int plugin_exmon_start(drakvuf_t drakvuf, const void *config) {
-    return 1;
-}
-static int plugin_exmon_stop(drakvuf_t drakvuf) {
-    return 1;
-}
-
-#endif
+        exmon(drakvuf_t drakvuf, const void *config);
+        ~exmon();
+};
 
 #endif
