+++ src/plugins/filedelete/filedelete.h
@@ -105,20 +105,28 @@
 #ifndef FILEDELETE_H
 #define FILEDELETE_H
 
-#ifdef ENABLE_PLUGIN_FILEDELETE
+#include "plugins/plugins.h"
 
-int plugin_filedelete_start(drakvuf_t drakvuf, const void *config);
-int plugin_filedelete_stop(drakvuf_t drakvuf);
+class filedelete: public plugin {
+    public:
+        drakvuf_trap_t traps[4] = {
+            [0 ... 3] = {
+                .lookup_type = LOOKUP_PID,
+                .u.pid = 4,
+                .addr_type = ADDR_RVA,
+                .type = BREAKPOINT,
+                .module = "ntoskrnl.exe",
+                .data = (void*)this
+            }
+        };
+        size_t* offsets;
 
-#else
-
-static int plugin_filedelete_start(drakvuf_t drakvuf, const void *config) {
-    return 1;
-}
-static int plugin_filedelete_stop(drakvuf_t drakvuf) {
-    return 1;
-}
-
-#endif
+        const char *dump_folder;
+        page_mode_t pm;
+        uint32_t domid;
+        output_format_t format;
+        filedelete(drakvuf_t drakvuf, const void *config);
+        ~filedelete();
+};
 
 #endif
