+++ src/plugins/objmon/objmon.cpp
@@ -102,36 +102,88 @@
  *                                                                         *
  ***************************************************************************/
 
-#include <stdarg.h>
-#include "plugins.h"
-#include "private.h"
+#include <config.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdbool.h>
+#include <string.h>
+#include <unistd.h>
+#include <sys/types.h>
+#include <sys/prctl.h>
+#include <sys/stat.h>
+#include <fcntl.h>
+#include <inttypes.h>
+#include <dirent.h>
+#include <glib.h>
+#include <err.h>
 
-int drakvuf_plugin_start(drakvuf_t drakvuf,
-                        drakvuf_plugin_t plugin,
-                        const void *config)
-{
-    if ( __DRAKVUF_PLUGIN_LIST_MAX != 0 &&
-         plugin < __DRAKVUF_PLUGIN_LIST_MAX)
+#include <libvmi/libvmi.h>
+#include "private.h"
+#include "objmon.h"
+
+/*
+ NTKERNELAPI
+ NTSTATUS
+ ObCreateObject (
+ IN KPROCESSOR_MODE ObjectAttributesAccessMode OPTIONAL,
+ IN POBJECT_TYPE ObjectType,
+ IN POBJECT_ATTRIBUTES ObjectAttributes OPTIONAL,
+ IN KPROCESSOR_MODE AccessMode,
+ IN PVOID Reserved,
+ IN ULONG ObjectSizeToAllocate,
+ IN ULONG PagedPoolCharge OPTIONAL,
+ IN ULONG NonPagedPoolCharge OPTIONAL,
+ OUT PVOID *Object
+ );
+ */
+static event_response_t cb(drakvuf_t drakvuf, drakvuf_trap_info_t *info) {
+
+    objmon *o = (objmon *)info->trap->data;
+    vmi_instance_t vmi = drakvuf_lock_and_get_vmi(drakvuf);
+    page_mode_t pm = vmi_get_page_mode(vmi);
+    uint8_t index = ~0;
+
+    access_context_t ctx;
+    ctx.translate_mechanism = VMI_TM_PROCESS_DTB;
+    ctx.dtb = info->regs->cr3;
+    ctx.addr = info->regs->rdx + o->typeindex_offset;
+
+    vmi_read_8(vmi, &ctx, &index);
+
+    if(index < WIN7_TYPEINDEX_LAST)
     {
-        return plugins[plugin].start(drakvuf, config);
+        switch(o->format) {
+        case OUTPUT_CSV:
+        {
+            printf("objmon,%s", win7_typeindex[index]);
+            break;
+        }
+        default:
+        case OUTPUT_DEFAULT:
+            printf("[OBJMON] %s", win7_typeindex[index]);
+            break;
+        };
+
+        printf("\n");
     }
 
+    drakvuf_release_vmi(drakvuf);
     return 0;
 }
 
-int drakvuf_plugins_stop(drakvuf_t drakvuf)
-{
-    int i;
-    int ret = 0;
+/* ----------------------------------------------------- */
 
-    if (__DRAKVUF_PLUGIN_LIST_MAX == 0)
-        return ret;
+objmon::objmon(drakvuf_t drakvuf, const void *config) {
+    const char *rekall_profile = (const char *)config;
 
-    for(i=0;i<__DRAKVUF_PLUGIN_LIST_MAX;i++) {
-        ret = plugins[i].stop(drakvuf);
-        if (!ret)
-            break;
-    }
+    if(VMI_FAILURE == drakvuf_get_function_rva(rekall_profile, "ObCreateObject", &this->trap.u2.rva))
+        return;
+    if (VMI_FAILURE==drakvuf_get_struct_member_rva(rekall_profile, "_OBJECT_HEADER", "TypeIndex", &this->typeindex_offset))
+        return;
 
-    return ret;
+    this->trap.cb = cb;
+    this->trap.data = (void*)this;
+
+    this->format = drakvuf_get_output_format(drakvuf);
+    drakvuf_add_trap(drakvuf, &this->trap);
 }
