+++ src/plugins/plugins.h
@@ -107,7 +107,17 @@
 
 #include <config.h>
 #include <stdlib.h>
-#include "../libdrakvuf/drakvuf.h"
+#include <libdrakvuf/libdrakvuf.h>
+
+/***************************************************************************/
+
+/* Plugin-specific configuration input */
+struct filedelete_config {
+    const char *rekall_profile;
+    const char *dump_folder;
+};
+
+/***************************************************************************/
 
 typedef enum drakvuf_plugin {
     PLUGIN_SYSCALLS,
@@ -119,15 +129,19 @@ typedef enum drakvuf_plugin {
     __DRAKVUF_PLUGIN_LIST_MAX
 } drakvuf_plugin_t;
 
-int drakvuf_plugin_start(drakvuf_t drakvuf,
-                        drakvuf_plugin_t plugin,
-                        const void *config);
-int drakvuf_plugins_stop(drakvuf_t drakvuf);
+class plugin {};
+class drakvuf_plugins
+{
+    private:
+        drakvuf_t drakvuf;
+        plugin* plugins[__DRAKVUF_PLUGIN_LIST_MAX] = { [0 ... __DRAKVUF_PLUGIN_LIST_MAX-1] = NULL };
 
-/* Plugin-specific configuration input */
-struct filedelete_config {
-    const char *rekall_profile;
-    const char *dump_folder;
+    public:
+        drakvuf_plugins(drakvuf_t drakvuf);
+        ~drakvuf_plugins();
+        int start(drakvuf_plugin_t plugin, const void* config);
 };
+
+/***************************************************************************/
 
 #endif
