+++ src/plugins/poolmon/poolmon.cpp
@@ -102,15 +102,6 @@
  *                                                                         *
  ***************************************************************************/
 
-/*
- * 1) ExAllocatePoolWithTag: is it a FILE?
- *   - YES: breakpoint RSP
- * 2) RSP: RAX -> _FILE_OBJECT
- *    - MEMTRAP W location
- * 3) MEMTRAP: is it at FileName string buffer?
- *    - YES: read string and remove trap
- */
-
 #include <config.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -129,265 +120,112 @@
 #include <libvmi/libvmi.h>
 #include "../plugins.h"
 #include "private.h"
+#include "poolmon.h"
 
-#define POOLTAG_FILE "Fil\xe5"
-#define ALIGN_SIZE(alignment, size) \
-    ( (size % alignment) ? (alignment - (size % alignment)) : 0 )
+static GTree* pooltag_build_tree() {
+    GTree *pooltags = g_tree_new((GCompareFunc) strcmp);
 
-static drakvuf_trap_t poolalloc;
-static GSList *writetraps;
-static GHashTable *rettraps;
-static addr_t file_object_size, file_name_offset,
-              string_buffer_offset, string_length_offset;
-static page_mode_t pm;
-static output_format_t format;
-
-struct rettrap_struct {
-    drakvuf_trap_t *trap;
-    long counter;
-};
-
-struct file_watch {
-    addr_t file_name_buffer;
-    addr_t file_name_length;
-};
-
-void free_writetrap(drakvuf_trap_t *trap) {
-    //printf("Freeing writetrap @ %p\n", trap);
-    writetraps = g_slist_remove(writetraps, trap);
-    free(trap);
-}
-
-static event_response_t file_name_cb(drakvuf_t drakvuf, drakvuf_trap_info_t *info) {
-    vmi_instance_t vmi = drakvuf_lock_and_get_vmi(drakvuf);
-    struct file_watch *watch = info->trap->data;
-
-    if (info->trap_pa == watch->file_name_buffer)
-    {
-        addr_t file_name = 0;
-        uint16_t length = 0;
-        vmi_read_addr_pa(vmi, watch->file_name_buffer, &file_name);
-        vmi_read_16_pa(vmi, watch->file_name_length, &length);
-
-        //printf("File name @ 0x%lx. Length: %u\n", file_name, length);
-
-        if (file_name && length > 0 && length < VMI_PS_4KB) {
-            unicode_string_t str = { .contents = NULL };
-            str.length = length;
-            str.encoding = "UTF-16";
-            str.contents = malloc(length);
-            vmi_read_va(vmi, file_name, 0, str.contents, length);
-            unicode_string_t str2 = { .contents = NULL };
-            status_t rc = vmi_convert_str_encoding(&str, &str2, "UTF-8");
-
-            if (VMI_SUCCESS == rc) {
-
-                switch(format) {
-                case OUTPUT_CSV:
-                    printf("filetracer,%u,%s\n", info->vcpu, str2.contents);
-                    break;
-                default:
-                case OUTPUT_DEFAULT:
-                    printf("[FILETRACER] VCPU:%u %s\n", info->vcpu, str2.contents);
-                    break;
-                };
-
-                g_free(str2.contents);
-            }
-
-            free(str.contents);
-            //printf("Requesting to free writetrap @ %p\n", info->trap);
-            drakvuf_remove_trap(drakvuf, info->trap, free_writetrap);
-        }
+    uint32_t i = 0;
+    for (; i < TAG_COUNT; i++) {
+        g_tree_insert(pooltags, (char*) tags[i].tag,
+                (char*) &tags[i]);
     }
 
-    drakvuf_release_vmi(drakvuf);
-    return 0;
-}
-
-/* This will be hit for all sorts of heap alloc returns */
-static event_response_t pool_alloc_return(drakvuf_t drakvuf, drakvuf_trap_info_t *info) {
-    vmi_instance_t vmi = drakvuf_lock_and_get_vmi(drakvuf);
-    struct rettrap_struct *s = info->trap->data;
-    addr_t obj_pa = vmi_pagetable_lookup(vmi, info->regs->cr3, info->regs->rax);
-    bool file_alloc = 0;
-    addr_t ph_base = 0, thread = 0;
-    uint32_t block_size = 0;
-    uint32_t tag;
-    uint32_t aligned_file_size = file_object_size;
-
-    if ( pm == VMI_PM_IA32E ) {
-        struct pool_header_x64 ph;
-        memset(&ph, 0, sizeof(struct pool_header_x64));
-        ph_base = obj_pa - sizeof(struct pool_header_x64);
-        vmi_read_pa(vmi, ph_base, &ph, sizeof(struct pool_header_x64));
-        block_size = ph.block_size * 0x10; // align it
-        if(!memcmp(&ph.pool_tag, &POOLTAG_FILE, 4))
-            file_alloc = 1;
-    } else {
-        struct pool_header_x86 ph;
-        memset(&ph, 0, sizeof(struct pool_header_x86));
-        ph_base = obj_pa - sizeof(struct pool_header_x86);
-        vmi_read_pa(vmi, ph_base, &ph, sizeof(struct pool_header_x86));
-        block_size = ph.block_size * 0x8; // align it
-        if(!memcmp(&ph.pool_tag, &POOLTAG_FILE, 4))
-            file_alloc = 1;
-    }
-
-    if (!file_alloc) goto done;
-
-    // We will need to catch when the file string buffer pointer is updated
-    addr_t file_base = ph_base + block_size - file_object_size; // addr of "_FILE_OBJECT"
-    addr_t file_name = file_base + file_name_offset; // addr of "_UNICODE_STRING"
-
-    struct file_watch *watch = g_malloc0(sizeof(struct file_watch));
-    watch->file_name_buffer = file_name + string_buffer_offset;
-    watch->file_name_length = file_name + string_length_offset;
-
-    //printf("PH 0x%lx. Block size: %u\n", ph_base, block_size);
-    //printf("File size: 0x%lx File base: 0x%lx. Unicode string @ 0x%lx. Last write @ 0x%lx\n",
-    //       file_object_size, file_base, file_name-file_base, watch->file_name_buffer-file_base);
-
-    drakvuf_trap_t *writetrap = g_malloc0(sizeof(drakvuf_trap_t));
-    //printf("Made a writetrap @ %p\n", writetrap);
-    writetrap->lookup_type = LOOKUP_NONE;
-    writetrap->addr_type = ADDR_PA;
-    writetrap->type = MEMACCESS_W;
-    writetrap->memaccess_type = POST;
-    writetrap->cb = file_name_cb;
-    writetrap->u2.addr = watch->file_name_buffer;
-    writetrap->data = watch;
-
-    writetraps = g_slist_prepend(writetraps, writetrap);
-
-    drakvuf_add_trap(drakvuf, writetrap);
-
-    s->counter--;
-
-    /*
-    if(s->counter == 0) {
-        drakvuf_remove_trap(drakvuf, info->trap, free);
-        g_hash_table_remove(rettraps, &s->trap->u2.addr);
-        free(s);
-    }*/
-
-done:
-    drakvuf_release_vmi(drakvuf);
-    return 0;
+    return pooltags;
 }
 
 static event_response_t cb(drakvuf_t drakvuf, drakvuf_trap_info_t *info) {
-
+    poolmon *p = (poolmon*)info->trap->data;
     vmi_instance_t vmi = drakvuf_lock_and_get_vmi(drakvuf);
-    reg_t tag = 0, size = 0;
+    page_mode_t pm = vmi_get_page_mode(vmi);
+    reg_t pool_type, size;
+    char tag[5] = { [0 ... 4] = '\0' };
 
-    access_context_t ctx = {
-        .translate_mechanism = VMI_TM_PROCESS_DTB,
-        .dtb = info->regs->cr3,
-    };
+    access_context_t ctx;
+    ctx.translate_mechanism = VMI_TM_PROCESS_DTB;
+    ctx.dtb = info->regs->cr3;
 
     if (pm == VMI_PM_IA32E) {
+        pool_type = info->regs->rcx;
         size = info->regs->rdx;
-        tag = info->regs->r8;
+        *(reg_t*)tag = info->regs->r8;
     } else {
+        ctx.addr = info->regs->rsp+12;
+        vmi_read_32(vmi, &ctx, (uint32_t*)tag);
         ctx.addr = info->regs->rsp+8;
         vmi_read_32(vmi, &ctx, (uint32_t*)&size);
-        ctx.addr = info->regs->rsp+12;
-        vmi_read_32(vmi, &ctx, (uint32_t*)&tag);
+        ctx.addr = info->regs->rsp+4;
+        vmi_read_32(vmi, &ctx, (uint32_t*)&pool_type);
     }
 
-    /*printf("Got a heap alloc request for tag %c%c%c%c!\n",
-           ((uint8_t*)&tag)[0],
-           ((uint8_t*)&tag)[1],
-           ((uint8_t*)&tag)[2],
-           ((uint8_t*)&tag)[3]
-    );*/
+    struct pooltag *s = (struct pooltag*)g_tree_lookup(p->pooltag_tree, tag);
 
-    if(!memcmp(&tag, &POOLTAG_FILE, 4)) {
-
-        addr_t ret, ret_pa;
-        ctx.addr = info->regs->rsp;
-        vmi_read_addr(vmi, &ctx, &ret);
-        ret_pa = vmi_pagetable_lookup(vmi, info->regs->cr3, ret);
-
-        struct rettrap_struct *s = g_hash_table_lookup(rettraps, &ret_pa);
-        if (s) {
-            s->counter++;
-        } else {
-            drakvuf_trap_t *rettrap = g_malloc0(sizeof(drakvuf_trap_t));
-            s = g_malloc0(sizeof(struct rettrap_struct));
-            s->trap = rettrap;
-            s->counter = 1;
-
-            rettrap->lookup_type = LOOKUP_NONE;
-            rettrap->addr_type = ADDR_PA;
-            rettrap->type = BREAKPOINT;
-            rettrap->name = "HeapRetTrap";
-            rettrap->cb = pool_alloc_return;
-            rettrap->u2.addr = ret_pa;
-            rettrap->data = s;
-
-            drakvuf_add_trap(drakvuf, rettrap);
-            g_hash_table_insert(rettraps, &rettrap->u2.addr, s);
-        }
-
-        //printf("File alloc request on vCPU %u. Ret: 0x%lx. Counter: %u\n",
-        //         info->vcpu, ret_pa, s->counter);
+    switch(p->format) {
+    case OUTPUT_CSV:
+    {
+        printf("poolmon,%s,%s,%zu", tag,
+               pool_type<MaxPoolType ? pool_types[pool_type] : "unknown_pool_type", (size_t)size);
+        if (s)
+            printf(",%s,%s", s->source, s->description);
+        break;
     }
+    default:
+    case OUTPUT_DEFAULT:
+        printf("[POOLMON] %s (type: %s, size: %zu)", tag,
+               pool_type<MaxPoolType ? pool_types[pool_type] : "unknown_pool_type", (size_t)size);
+        if (s)
+            printf(": %s,%s", s->source, s->description);
+
+        break;
+    };
+
+    printf("\n");
 
     drakvuf_release_vmi(drakvuf);
     return 0;
 }
 
+/*void pool_tracker_free(vmi_instance_t vmi, vmi_event_t *event) {
+    // TODO: 32-bit inputs are on the stack
+    reg_t rcx, rdx;
+    drakvuf_t *drakvuf = event->data;
+
+    vmi_get_vcpureg(vmi, &rcx, RCX, event->vcpu_id);
+    vmi_get_vcpureg(vmi, &rdx, RDX, event->vcpu_id);
+
+    char ctag[5] = { [0 ... 4] = '\0' };
+    memcpy(ctag, &rdx, 4);
+
+    PRINT(drakvuf, HEAPFREE_STRING, rcx, ctag);
+}*/
+
 /* ----------------------------------------------------- */
 
-int plugin_filetracer_start(drakvuf_t drakvuf, const char *rekall_profile) {
-    vmi_instance_t vmi = drakvuf_lock_and_get_vmi(drakvuf);
-    pm = vmi_get_page_mode(vmi);
-    drakvuf_release_vmi(drakvuf);
-    rettraps = g_hash_table_new(g_int64_hash, g_int64_equal);
-    format = drakvuf_get_output_format(drakvuf);
+poolmon::poolmon(drakvuf_t drakvuf, const void *config) {
+    const char *rekall_profile = (const char*)config;
+    this->pooltag_tree = pooltag_build_tree();
 
-    poolalloc.lookup_type = LOOKUP_PID;
-    poolalloc.u.pid = 4;
-    poolalloc.addr_type = ADDR_RVA;
-    poolalloc.name = "ExAllocatePoolWithTag";
-    poolalloc.module = "ntoskrnl.exe";
-    poolalloc.type = BREAKPOINT;
-    poolalloc.cb = cb;
+    this->trap.lookup_type = LOOKUP_PID;
+    this->trap.u.pid = 4;
+    this->trap.addr_type = ADDR_RVA;
+    if (VMI_FAILURE == drakvuf_get_function_rva(rekall_profile,
+                                                "ExAllocatePoolWithTag",
+                                                &this->trap.u2.rva))
+    {
+        return;
+    }
 
-    if (VMI_FAILURE == drakvuf_get_function_rva(rekall_profile, "ExAllocatePoolWithTag", &poolalloc.u2.rva))
-        return 0;
-    if (VMI_FAILURE == drakvuf_get_struct_member_rva(rekall_profile, "_FILE_OBJECT", "FileName", &file_name_offset))
-        return 0;
-    if (VMI_FAILURE == drakvuf_get_struct_member_rva(rekall_profile, "_UNICODE_STRING", "Buffer", &string_buffer_offset))
-        return 0;
-    if (VMI_FAILURE == drakvuf_get_struct_member_rva(rekall_profile, "_UNICODE_STRING", "Length", &string_length_offset))
-        return 0;
-    if (VMI_FAILURE == drakvuf_get_struct_size(rekall_profile, "_FILE_OBJECT", &file_object_size))
-        return 0;
+    this->trap.name = "ExAllocatePoolWithTag";
+    this->trap.module = "ntoskrnl.exe";
+    this->trap.type = BREAKPOINT;
+    this->trap.cb = cb;
+    this->trap.data = (void*)this;
+    this->format = drakvuf_get_output_format(drakvuf);
 
-    if (pm == VMI_PM_IA32E)
-        file_object_size += ALIGN_SIZE(16, file_object_size);
-    else
-        file_object_size += ALIGN_SIZE(8, file_object_size);
-
-    drakvuf_add_trap(drakvuf, &poolalloc);
-
-    return 1;
+    drakvuf_add_trap(drakvuf, &this->trap);
 }
 
-int plugin_filetracer_stop(drakvuf_t drakvuf) {
-
-    GSList *loop = writetraps;
-    while(loop) {
-        free(loop->data);
-        loop=loop->next;
-    }
-    g_slist_free(writetraps);
-
-    if ( rettraps )
-        g_hash_table_destroy(rettraps);
-
-    return 1;
+poolmon::~poolmon() {
+    if(this->pooltag_tree)
+        g_tree_destroy(this->pooltag_tree);
 }
